requirements.txt
requests>=2.32.0
streamlit>=1.30.0
matplotlib>=3.8.0
docxtpl>=0.16.7
python-docx>=1.1.0
pytest>=8.0.0
